requirements.txt
logging
colorlog
flask
psutil
requests
mysql.connector
plotly